requirements.txt
httpx
beautifulsoup4
lxml
python-dotenv

